feat: open logger from action
