+++ desktop/bridge/logger.ts
@@ -25,3 +25,5 @@ export interface LogEntry {
 export const logStorage = createChannelBridge<LogEntry>("log-storage");
 
 export const getAllLogsBridge = createInvokeBridge<void, LogEntry[]>("get-all-logs");
+
+export const requestToggleLoggerWindow = createInvokeBridge<void, void>("toggle-logger-window");
