+++ desktop/electron/appState.ts
@@ -1,5 +1,5 @@
 import { BrowserWindow } from "electron";
-import { makeObservable, observable } from "mobx";
+import { autorun, makeObservable, observable, runInAction } from "mobx";
 
 import { appWindowValue } from "@aca/desktop/bridge/appWindow";
 import { autorunEffect } from "@aca/shared/mobx/utils";
@@ -7,9 +7,11 @@ import { autorunEffect } from "@aca/shared/mobx/utils";
 export const appState = makeObservable(
   {
     mainWindow: null as null | BrowserWindow,
+    loggerWindow: null as null | BrowserWindow,
   },
   {
     mainWindow: observable.ref,
+    loggerWindow: observable.ref,
   }
 );
 
@@ -36,5 +38,23 @@ autorunEffect(() => {
     mainWindow.off("focus", handleFocus);
 
     mainWindow.off("blur", handleBlur);
+  };
+});
+
+autorun(() => {
+  const { loggerWindow } = appState;
+
+  if (!loggerWindow) return;
+
+  const removeLoggerWindow = () => {
+    runInAction(() => {
+      appState.loggerWindow = null;
+    });
+  };
+
+  loggerWindow.on("closed", removeLoggerWindow);
+
+  return () => {
+    loggerWindow?.off("closed", removeLoggerWindow);
   };
 });
