+++ desktop/envData.ts
@@ -1,5 +1,8 @@
+type WindowName = "Root" | "Logger";
+
 export interface AppEnvData {
   version: string;
   isDev: boolean;
   sentryDsn: string;
+  windowName?: WindowName;
 }
